requirements.txt
twikit
google-genai
flask
